Fix vertices.py
Again, changed dtype of Dipole and Charge fields.

+++ icenumerics/vertices.py
@@ -96,8 +96,8 @@ class vertices():
                           ('Location', float,(2,)),
                           ('id',int),
                           ('Coordination',int),
-                          ('Charge',int),
-                          ('Dipole',int,(2,))])
+                          ('Charge',float),
+                          ('Dipole',float,(2,))])
         
     def colloids_to_vertices(self,C):
